# quotapool: add fast path to Acquire