In recent admission control work it was determined that the inefficiency of the
acquisition path was artificially limiting throughput. This change is
especially critical to make the common case where there is a large amount of
traffic but generally enough quota. When applied to a current admission control
prototype it moves changes the impact of the read quota from having a ~30%
overhead on peak throughput of KV95 on 16-core machines to not having any
affect on peak throughput and not or detecting overload until the system
reaches a concurrency where throughput has already fallen by over 10%.

```
name                                            old time/op    new time/op    delta
ConcurrentIntQuotaPool/workers=512,quota=511-8    3.76µs ±28%    2.14µs ±70%  -43.12%  (p=0.001 n=10+10)
```

```
name                                            old time/op    new time/op    delta
IntQuotaPool-8                                     625ns ± 9%     238ns ±11%   -61.97%  (p=0.000 n=10+9)
ConcurrentIntQuotaPool/workers=1,quota=1-8        1.08µs ± 1%    0.37µs ± 5%   -65.73%  (p=0.000 n=8+10)
ConcurrentIntQuotaPool/workers=2,quota=2-8        1.49µs ±19%    0.42µs ± 5%   -72.08%  (p=0.000 n=10+9)
ConcurrentIntQuotaPool/workers=8,quota=4-8        2.69µs ± 1%    2.73µs ± 8%      ~     (p=1.000 n=7+10)
ConcurrentIntQuotaPool/workers=128,quota=4-8      3.43µs ±50%    2.76µs ±20%   -19.58%  (p=0.019 n=10+10)
ConcurrentIntQuotaPool/workers=512,quota=128-8    3.13µs ±24%    3.35µs ±53%      ~     (p=0.842 n=9+10)
ConcurrentIntQuotaPool/workers=512,quota=513-8    4.00µs ±37%    0.87µs ± 4%   -78.31%  (p=0.000 n=10+9)
IntQuotaPoolFunc-8                                1.02µs ± 6%    0.32µs ± 5%   -68.58%  (p=0.000 n=8+9)

name                                            old alloc/op   new alloc/op   delta
IntQuotaPool-8                                     0.00B          0.00B           ~     (all equal)
ConcurrentIntQuotaPool/workers=1,quota=1-8         0.00B          0.00B           ~     (all equal)
ConcurrentIntQuotaPool/workers=2,quota=2-8         0.00B          0.00B           ~     (all equal)
ConcurrentIntQuotaPool/workers=8,quota=4-8         0.00B          0.00B           ~     (all equal)
ConcurrentIntQuotaPool/workers=128,quota=4-8       0.00B          0.00B           ~     (all equal)
ConcurrentIntQuotaPool/workers=512,quota=128-8     1.00B ± 0%     1.00B ± 0%      ~     (all equal)
ConcurrentIntQuotaPool/workers=512,quota=513-8     3.60B ±94%     0.00B       -100.00%  (p=0.000 n=10+10)
IntQuotaPoolFunc-8                                 16.0B ± 0%     16.0B ± 0%      ~     (all equal)

name                                            old allocs/op  new allocs/op  delta
IntQuotaPool-8                                      0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=1,quota=1-8          0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=2,quota=2-8          0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=8,quota=4-8          0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=128,quota=4-8        0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=512,quota=128-8      0.00           0.00           ~     (all equal)
ConcurrentIntQuotaPool/workers=512,quota=513-8      0.00           0.00           ~     (all equal)
IntQuotaPoolFunc-8                                  1.00 ± 0%      1.00 ± 0%      ~     (all equal)
```

Release note: None