@ajwerner 2019-07-09
@andreimatei tagging you on this review because you're a little bit more familiar with the code. This implementation is both more efficient and I think clearer. I tagged @nvanbenschoten mostly because we'd been talking about it just before I typed it. No rush and thanks!