@ajwerner 2019-07-26
> Just spectating on this code and noticed this. sync.Pools are somewhat heavyweight. When a pool is first used after a GC, it is put on a global list of all sync.Pools which is then iterated over during GC in a world-is-stopped phase. Each GC clears the global list of pools requiring a pool to be re-added. Having a few dozen pools in CRDB seems reasonable, but if we were to have a QuotaPool per range I think we'd see problems. Am I missing something here?

Good catch! Fortunately we haven't adopted this in storage just yet (PR draft exists but I want to fix this quota pool leak first). See https://github.com/cockroachdb/cockroach/pull/39124.